+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.11.0-dev
 
 - Enhancements
+    - All configured databases can be reloaded in the background
+      using `Geolix.reload_databases/0`
     - Database configuration has been extended to support upcoming adapters
     - "GeoIP-Enterprise" databases return structs instead of plain maps
     - Lookups are done with a configurable timeout